Added dropdown menus to each series list row
The two menu items we need are in place but don't do anything yet.

For the glyphicons-option-vertical labal, found a problem with the yesod
scaffold. It was missing the glyphicons-halflings-regular.woff2 file.
Retrieved this from the bootstrap distribution zip file and added to the
project.

Also, made this work similarly to Google Play Music where the menu icon
only shows up when the row is hovered, so it's less visually noisy.